requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
alembic==1.13.2
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
python-dotenv==1.1.1
Pillow==10.4.0

aiogram==3.22.0
aiohttp==3.12.15
redis==5.0.8
psycopg2-binary==2.9.9
httpx==0.28.1
pytest==8.3.2
